- Changed: GameObject.Realm has now changed from a byte to an eRealm (which can still be cast to a byte).
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@954 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/gameobjects/GameMerchant.cs
@@ -350,7 +350,7 @@ namespace DOL.GS
 			merchant.Heading = Heading;
 			merchant.Speed = MaxSpeedBase;
 			merchant.Region = CurrentRegionID;
-			merchant.Realm = Realm;
+			merchant.Realm = (byte)Realm;
 			merchant.Model = Model;
 			merchant.Size = Size;
 			merchant.Level = Level;
